refactor: extract dashboard columns to dedicated hook
Moves column definitions from DeprecationsDashboard component into a
useDeprecationDashboardColumns hook to match codebase conventions and
resolve SonarQube nested component warning.
